% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gg_resX}
\alias{gg_resX}
\title{Generate residual plot of residuals against predictors}
\usage{
gg_resX(lm_object)
}
\arguments{
\item{lm_object}{lm object that contains fitted regression}
}
\value{
A list of ggplot objects that contains residual plot
of residuals against predictor values
}
\description{
Generate residual plot of residuals against predictors
}
\examples{
gg_resX()

}

