% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all.R
\name{plot_all}
\alias{plot_all}
\title{Plot all given plots in a square matrix form.}
\usage{
plot_all(plots, ncol = NULL)
}
\arguments{
\item{plots}{a list of plots}
}
\value{
plots in a given list arrangeed using gridExtra
}
\description{
Plot all given plots in a square matrix form.
}

