% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_boxcox.R
\name{gg_boxcox}
\alias{gg_boxcox}
\title{Plot boxcox graph in ggplot with suggested lambda transformation}
\usage{
gg_boxcox(lm_object, showlambda = TRUE, lambdaSF = 3)
}
\arguments{
\item{showlambda}{logical which controls whether lambda value should be displayed on graph. Defaults to TRUE}

\item{lambdaSF}{controls to how many significant figure is lambda rounded to. Defaults to 3.}

\item{lm}{lm object that contains fitted regression}
}
\value{
A ggplot object that contains boxcox graph
}
\description{
Plot boxcox graph in ggplot with suggested lambda transformation
}
\examples{
library(MASS)
data(Cars93)
cars_lm <- lm(Rev.per.mile ~ Passengers + Length + RPM, data = Cars93)
gg_boxcox(cars_lm)

}

