% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gg_qqplot}
\alias{gg_qqplot}
\title{Plot quantile-quantile plot (QQPlot) in ggplot with qqline shown.}
\usage{
gg_qqplot(lm_object)
}
\arguments{
\item{lm}{lm object that contains regression}
}
\value{
A qqplot with fitted qqline
}
\description{
Plot quantile-quantile plot (QQPlot) in ggplot with qqline shown.
}
\examples{
gg_qqplot*()
}

