% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_resfitted.R
\name{gg_resfitted}
\alias{gg_resfitted}
\title{Generate residual plot of residuals against fitted value}
\usage{
gg_resfitted(lm_object)
}
\arguments{
\item{lm}{lm object that contains a fitted regression}
}
\value{
A ggplot object
}
\description{
Generate residual plot of residuals against fitted value
}
\examples{
library(MASS)
data(Cars93)
cars_lm <- lm(Rev.per.mile ~ Passengers + Length + RPM, data = Cars93)
gg_resfitted(cars_lm)

}

