% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_resX.R, R/gg_resX_fortify.R
\name{gg_resX}
\alias{gg_resX}
\title{Generate residual plot of residuals against predictors}
\usage{
gg_resX(lm_object, plotAll = TRUE)

gg_resX(lm_object, plotAll = TRUE)
}
\arguments{
\item{lm_object}{a fitted linear model (i.e. lm, glm) that contains fitted regression}

\item{plotAll}{boolean value to determine whether plot will be return as 
a plot arranged using `grid.arrange()`. When set to false, the function
would return a list of residual plots. Parameter defaults to TRUE.}

\item{fitted.lm}{a fitted linear model (i.e. lm, glm) that contains fitted regression}

\item{data}{original dataset}

\item{plotAll}{boolean value to determine whether plot will be return as 
a plot arranged using `grid.arrange()`. When set to false, the function
would return a list of residual plots. Parameter defaults to TRUE.}
}
\value{
An arranged grid of residuals against predictor values plots in ggplot. 
If plotall is set to FALSE,  a list of ggplot objects will be returned instead. 
Name of the plots are set to respective variable names.

A list of ggplot objects that contains residual plot
of residuals against predictor values
}
\description{
Generate residual plot of residuals against predictors

Generate residual plot of residuals against predictors
}
\examples{
library(MASS)
data(Cars93)
# a regression with categorical variable
cars_lm <- lm(Price ~ Passengers + Length + RPM + Origin, data = Cars93)
gg_resX(cars_lm, data = Cars93)
# customize which diagnostic plot is included by have gg_resX to return a list of plots
plots <- gg_resX(cars_lm, data = Cars93, plotAll = FALSE)
names(plots)     # get name of the plots
exclude_plots <- plots[-1 ]    #exclude certain residual plots
include_plots <- plots[1]      # include certain residual plots
plot_all(exclude_plots)              # make use of plot_all() in lindia
plot_all(include_plots)
gg_resX()
library(MASS)
data(Cars93)
# a regression with categorical variable
cars_lm <- lm(Rev.per.mile ~ Passengers + Length + RPM + Origin, data = Cars93)
gg_resX(cars_lm, data = Cars93)
# customize which diagnostic plot is included by have gg_resX to return a list of plots
plots <- gg_resX(cars_lm, data = Cars93, plotAll = FALSE)
names(plots)     # get name of the plots
exclude_plots <- plots[-1 ]    #exclude certain residual plots
include_plots <- plots[1]      # include certain residual plots
plot_all(exclude_plots)              # make use of plot_all() in lindia
plot_all(include_plots)
}

